Fixed a race-condition in DS485Proxy::execute
Fixes #197

FrameBucketBase handled it's registration/de-registration in it's
constructor/destructor. This was no problem until we've introduced
a base-class that made FrameBucketBase::addFrame pure virtual.

If a descendant of FrameBucketBase gets destoyed, it will remain
in the bucket list of the proxy until FrameBucketBases destructor
got called. During that short period the entry of addFrame in the
vmt is of that object is null, as the descendant's destructor has
already been called.

I've taken out the automagic registration and replaced with manual
registration/de-registration. To ensure that allocated buckets get
removed from the bucket list I've added a static function that
can be passed to a boost::shared_ptr.