Michael Tross: Handling of dS485 bus reorg.
Communication failures on the RS485 interface may occur as devices are
powered on and off, leading to lost packets. In some circumstances this
may lead to a reorganization of the bus where all devices have to run
through the initial address allocation phase. The dSS does not detect or
properly handle this bus reorganization.

Furthermore, during this phase the token is lost and therefore no API
level communication can take place. This will lead to timeouts of
requests issued by the data model and subsequent retransmissions or
false interpreation of a device failure.

Closes #254