Update cmake files
The most important change is using find_package for finding all the
dependencies. Find*.cmake files are added under cmake/Modules/. It only
checks for the Mozilla SpiderMonkey JavaScript library. Don't know
whether this is sufficient. It is tested on Debian GNU Linux. Don't know
whether it works on Darwin.

The changes include:

 * Change to minimum cmake version 2.8

 * Change cmake functions to lower case letters

 * Using find_package to search for packages in the system

 * Add Boost as dependency using find_package