Configured GSOAP to use SSL
This commit implements the necessary changes to enable SSL with GSOAP.

Unfortunately GSOAP did not properly generate C++ code to support SSL,
I had to subclass the generated dssService object and extend it with the
required ssl_server_context function. This should be kept in mind when
regenerating the soap files, if something changes in the future it may be
possible to adapt the dssSSLService subclass.

- cmake updated to look for gsoapssl++ and openssl libraries
- webservices extended to use soap's ssl functions
- wsdl updated to point to the https url
- urls in example python scripts updated to point to the https url

references #168