Support for removing non present devices from JSON
This commit adds a "removeDevice" function to the JSON API, the device
must be specified by the devID parameter.

Reorganized the StructureRequestHandler class a bit, put contents of the
huge if block into separate functions (i.e. function per request) and
added the removeDevice request processing.

Added documentation to WebServerAPI, the whole structure stuff was not
yet documented there.

references #292