First version of cookie based session handling
All json calls will now look for a session cookie "token" and will ask
the session manager for the session. If the cookie is not found, a
Set-Cookie header with an appropriate session id is generated and passed
to the browser. The browser will then issue all further request with our
cookie in the headers, which allows us to identify the session.

The session will be passed to the handlers, which may store any data
they need.

TODO
* currently sessions remain forever, we need a mechanism to track idle
  sessions and remove them after they time out

* so far the cookies are handled in the json callback, we need to add
  some sort of "cookie manager" class that will take care of parsing the
  cookie, generating a new session and updating the cookie, etc.

* make sure that the cookies are handled not only within the json
  callback, but for all HTTP requests going to dSS