Localtime <-> UTC setup
Added a way to remember the local time offsets before we switch to the
UTC timezone. This will allow us to correctly convert UTC to local time
and also to convert local time input to UTC.

A possibility to solve this would be to write an own singleton class that would
get initialized with the correct offset in the main function, however I
considered this an overhead and went for an static function in the DateTime
class which initializes a static variable.

Note, that on non linux system the conversion will not account for
daylight savings and might be off (this is how the existing
implementation was handling it).

Added code for linux systems, that will use tm_gmtoff which does deliver
the correct offset accounting for daylight savings and thus ensures taht
the conversion is correct.