Seconds since epoch must be in UTC
Turned out that JavaScript Date objects that are created from epoch
counter are always in UTC. If we were to return local time here, one
would have to recalculate that manually in JS to generate a correct Date
object.

So we will return UTC here and document this accordingly, in order to
make life of JS people a little easier.