Made sure dSM presence is updated correctly
Made use of the library callbacks to ensure that we set the presence
flag for dSMs correctly.

Removed ModelEvent::etNewDSMeter event which became obsolete, adding of
new ds485 devices is handled by ModelEvent::etDS485DeviceDiscovered
which was introduced earlier.

Made sure that presence flag of devices attached to a leaving dSM is
also updated correctly.

Fixes #477, #462