Added return code check function for async calls
Calls to libdsm that are normally synchronious and return results
immediately behave differently when the broadcast address is used. In
this case the function will return -2, this indicates that the output
parameters that normally carry the result are invalid, and that the
results will arrive asynchroniously.

Added an error check function that will account for this library
behavior.

Also made sure that we print the return code value for unhandled error
codes.