Handle case where no one is logged in
Sometimes it comes to a situation where a session exists, but noone is
logged in. As a result such a session will never authenticate, even if
the request is coming from the trusted port. When this happens, we will
try to get the user from the HTTP request, just as it is done for new
sessions.

implements #1099