Javascript profiler jumbo patch.
Will measure invocations of a javascript handler with a convenient
StopWatch wrapper class. This will only do the minimal work, such as
taking a timestamp and calculationg the delta between two timestamps.
Such a measurement is added to global datastructure, which is regularly
sorted and update to the property tree through the means of the
BenchmarkPubplisherPlugin.

Signed-off-by: Andreas Fenkart <andreas.fenkart@dev.digitalstrom.org>