security: remove Privilege::Read since all privileges have it anyway.
Th user/system roles have it anyway and the fallback privilege for
non-authenticated users, aka nobody, has it too. So the only way to
fail this test is to create a new role assign a privilege to it, and
not giving it read access. This is so remotely useful so removing the
whole feature.
Checking read access is not cheap, because privileges are only set for
the top '/' and the '/system/security' node. So when checking read
access for a property we first have to recurse until we meet either
of these nodes.
UNIT TEST CHANGES: it feels natural that authenticated users have the
same read access rights, as the unauthenticated 'Nobody' user.

Signed-off-by: Andreas Fenkart <andreas.fenkart@dev.digitalstrom.org>