util: refactor getRandomSalt function
drop the reinterpret_cast<long long int*>, it was probably
used to convert the byte stream to hex in one line. which worked
but is quite unreadable.
Not sure why the salt has to be long long int, probably because it's
the longest pointer available
NOTE: non-linux platforms will not compile after this change. Also
the DSS will abort on startup if urandom is not found on host system.

Signed-off-by: Andreas Fenkart <andreas.fenkart@dev.digitalstrom.org>