webservice: caller can install hook to receive outcome of call.
Current webservice calls are best effort, in case of network failure
no reattempts happen. The current interface does not allow to start
retries, since there is no hook to implement such retransmit logic.
This commit adds a hook, it is planned to be used for unit testing,
where we want to check the reply against some known values

Signed-off-by: Andreas Fenkart <andreas.fenkart@dev.digitalstrom.org>