scripts: convert subscriptions timing dump into sorted list.
The 'system/js/timings' tree is uploaded to statistics twice a day
via the logfile uploader. the content uploaded is a nested json
structure created via dss_json.sh.
This script takes such a dump as input and extracts subscription/count
and total time and prints it in sorted order.

Signed-off-by: Andreas Fenkart <andreas.fenkart@dev.digitalstrom.org>