web: session: store subscription id in class
subscription id is 1:1 connected with EventSubscriptionSession instance.
By adding the storing the subscription id with the instance, we can use
a simple container such as vector/list and use std::find_if to select
the right instance.
Currently we use std::map which is more expensive for adding/removal
of elements.

Signed-off-by: Andreas Fenkart <andreas.fenkart@dev.digitalstrom.org>