Check sensor timestamps every 10 minutes and mark sensor data as invalid
If lifetime of the sensor data is exceeded, the sensor data is marked as
invalid. Additionally, an event is fired.

Implementation notes:
although it is possible to check the validity per individual sensor, the
global sensor data flag of the device which is available in JSON will
only have the "valid" state if all sensors of the device are valid.

The first sensor check is started after the model_ready event,
afterwards its re-queued every 10 minutes.

implements #6471