base: splitString simplify skip '\' handling
if the delimiter key is prefixed with '\' then the ignore the delimiter.
treat it as a regular character. But we do unescape it, which means
we remove the '\' and continue search for the next delimiter following
its position.
Caution: now that the delimiter is no more escaped so we need restart
the search after that character

Signed-off-by: Andreas Fenkart <andreas.fenkart@dev.digitalstrom.org>