Fix dSM discovery
Sometimes the buschange callback will arrive with a dSM that is not yet
known to the model. In this case the dSM device type will also not be
known - in this case the old code just ignores it.

Instead, when we get an unknown dSM - send an event to the model and
force a rescan.