Do not rely on invalid data from persistent file
ref #8530

A new attribute for the ds485 bus members was introduced for 1.9.x.
When updating a dss from 1.8.3 to 1.9.x, the old apartment.xml is
read with no bus member type set. (the type is set to BusMember_Unknown)
At startup, it can happen that the function handleBusChange is called
and the type is read from the invalid persistent data.
The 'invalid device' will then be ignored and will remain in state
'not present' and 'offline'.