EventUpload: abstract Uploader interface instaed of C callback
For unit tests need to bind the callback to an object.
Typical c callbacks have a priv NULL pointer to sneak in the object.
For C++ this is very odd, and passing NULL when creating shared object
see make_shared<SensorLog>("mshub, .., NULL)
Other options was to make uploader a class template of SensorLog, which
I assume leads to code bloat

Signed-off-by: Andreas Fenkart <andreas.fenkart@dev.digitalstrom.org>