Check if resetting of the rrd database is needed
ref #7575

If the time difference between the last entry in the rrd database and the new data is bigger than 1 Week, the database will be cleansed.
If the time difference is between a 1 day and a week, the data is written to the rrd database and a flush command is initiated to synchronize data.
Instead of checking db consistency when getLastEnergyCounter is triggered,
the db consistency is checked periodically. The getLastEnergyCounter is relying on that.