UMR: reset buttonInputMode to STANDARD when changing color from Joker
On a "second dsid", where no 2way configuration setting is possible,
the buttonInputMode would not be reset when switching from BinarySensor
mode to ButtonMode mode.

This change will potentially lead to an "unnecessary" parameter write to
the terminal block when the real buttonInputMode is written directly after
changing the color.

refs #9395