dsuid: overload ==/!= operators for dsuid_t
dsuid_equal does not work with temporary values see:
dsid_equal(&dsm->getDSID(), &otherDSID)
that does not work since getDSID returns the dsuid on the stack and
taking the address of that value to pass it to dsuid_equal, is dangerous
if the function kept that pointer after exit.
operator==(dsuid &lhs, dsuid &rhs) which is a thin wrapper around
dsuid_equal. But since it uses *const* references, it feels odd if the
function kept a reference after exit and the compiler does not emit a
warning

Signed-off-by: Andreas Fenkart <andreas.fenkart@dev.digitalstrom.org>