tests: add DSSInstanceFixture to SensorUploaderTest
Sensor uploader exports counters to the property system. This will
call getInstance, hence create a an instance implicitly.
This instance is neither properly configured nor destroyed at the end,
If followup tests rely on hasInstance to be false, they will can fail
since hasInstance is true but the instance not fully functional

Signed-off-by: Andreas Fenkart <andreas.fenkart@dev.digitalstrom.org>