handler: state: add lookupState that return true/false if state has been found
typical use case is: if-state-exists-then-do-something. With exceptions
exceptions we don't need the check, but need to wrap the code with
try/catch clause, to prevent program abort.
Since lookupState is thread safe, we can drop the catch, and by plugging
the lookupState function into the if statement, we can squeeze lookup
state and check if it exists into one line.
The catch clause is a saving if lots of functions can throw (the same)
exception, e.g. getZone/getGroup. But for lookup up states it means
3 lines (try-fetch-catch-closing-brace-on-same-line) down to one
Mind that there is no error handling in case the state/zone/group does
not exist, so if they don't exist something is probably odd anyway,
but since we have javascript, which we do not fully trust this might
happen

Signed-off-by: Andreas Fenkart <andreas.fenkart@dev.digitalstrom.org>