handler: states: pass state to stateBinaryInputGeneric
The stateBinaryInputGeneric function is looking up the state as the
first step. If the state does not exist, an exception is thrown, caught
and the function simply returns.
getOrCreateState is typically called just before stateBinaryInputGeneric.
The former returns the state, but was ignored.
Regarding wind and rain: we only create the state if the group
is bigger than AppUserGroupMin. But there are also global rain/wind
states that are created at boostrap. We need to update those as well.

Signed-off-by: Andreas Fenkart <andreas.fenkart@dev.digitalstrom.org>