model: main: do not erase events from event queue
it creates jumps in the processed events counter, which leads to
nasty side effects. delay flushing the apartment.xml to disk by
30s which also combines multiple dirty events, and prevents not
writing apartment due to ever new arriving dirty events.
Regarding ApartmentChange notifications we mimick the existing behaviour
by emitting the event upon the first event, but all currently queue
etModelDirty will not signal. This is equivalent to actually erasing
them from the queue. Problem is that webservice(cloud) needs to be
notified in timely fashion, but it is not reading the apartment.xml so
that is an independent problem.

Signed-off-by: Andreas Fenkart <andreas.fenkart@dev.digitalstrom.org>