Revert "web: drain ds485 model changes before responding"
This reverts commit 05745bbd83021b13f99f5e298f3f45390036f912.
Whenever the busscanner runs, which happens each time a
DSM joins/leaves, the modelmaintenance thread will nor process
model events.
This is extremely annoying on startup, since the administration
UI complains about the DSS not running, which is running, just
not responding to the web requests, since they are stucked in
the model barrier.
Also on dss11 sd-card, with legacy jffs2, this is quiet annoying
since the ds485d regularly hangs and loses bus conection

Signed-off <andreas.fenkart@digitalSTROM.org>