http client: drop unneded shared_ptr from header argument type
the header are copied to curl_slist before passing off to the curl
library. passing the headers by reference is enough to keep the headers
valid while referencing them
in the webservice connectin we rely on the copy-construct of HashMap
to copy them into the interim HttpRequest structure. STL containers
do have copy-contructors, the string-elements are copied as well, not
passed by reference

Signed-off-by: Andreas Fenkart <andreas.fenkart@dev.digitalstrom.org>