Fix thanos visibility issues
Saving visibility and pired devices data to apartment.xml made no sense,
because setting of those properties on the device class depends on the
knowledge if the device in question is actually a TNY.

Saing the product ID did not help either, because the order in the XML
is random and there is no guarantee that the product ID will be read out
before the visibility attributes.

Following change is made to work around the above problems:

by default all devices are visible. Once the TNY config reader kicks in,
it will set the actual visibility attributes which will result in the
correct property tree update.

implements #11764