Do not invalidate already loaded OEM product information
When internet access was not available, all OEM product info got
invalidated and removed, eventhough it may have been correct because of
an earlier readout. As one of a side effects icons disappeared, etc.

Now we will remember the OEM product state before the readout and
restore the old state in case the readout fails. So things that were
unknown, stay unknown, but data that was valid stays valid.

implements #12912