set vdcs in idle state.
refs #13176

If bus member type is known and type is vdc, set it to idle state.
Retry, if meter is not synchronized and bus member type not available.
(This can happen, if a new vdc appears and there
are no data in persistent storage. Returning false in this case means,
that the datamodel is incomplete and that we should retry later)