Save number of paired TNYs in apartment.xml
Sometimes the dSS does not know the number of paired devices anymore
(for instance when the device was inactive upon dSS restart). In these
cases removing the first dSUID did not perform the cleanup of paired
dSUIDs, because the dSS did not have the pairing information anymore.

We will now save the pairing info persistently to close this loophole.

Implements #13524