dss: eventQ: setup the event queue references early
interpreter/queue/runner all talk to each other directly.
e.g. pushing an event to the (ready-) queue will forward the
event to the (peinding-) runner. The same happens when using the
pushEvent function of the interpreter.
For unit tests to work without starting the main dss thread, these
references must be present, or it segfaults.

Signed-off-by: Andreas Fenkart <andreas.fenkart@dev.digitalstrom.org>