eventtrigger: raise event with configurable lag and rescheduling
If a trigger matches an event does not have to be raised now, but
with a configurable lag. If the trigger matches again, the pending
event can be rescheduled so that it executes in now + delay seconds.
Runtime nodes (eventid + timestamp) are created under triggerPath +
"lag/" same place as the configuration parameters

Signed-off-by: Andreas Fenkart <andreas.fenkart@dev.digitalstrom.org>