taskprocesser: pass TaskProcessor instance by reference
passing the instance as a shared_ptr is futeless, since the
TaskProcess is created at program startup and never destroyed.
It could also be modeled as a singleton.
Made TaskProcessor copy constructor private wihout implementing
it, to prevent accidentally copying the instance
Evtl. make it real singleton or move getter to dSS class.

Signed-off-by: Andreas Fenkart <andreas.fenkart@dev.digitalstrom.org>