db download: rename property to "id" to make it more generic
non-script entities will be processing this event too. Rename to
"id" since that is more generic. It is free chosable string by
the party interested in receiving a notification that download
completed failed or not.
Also adding url to the result event, since that is useful if several
db need to be maintained under the same id

Signed-off-by: Andreas Fenkart <andreas.fenkart@dev.digitalstrom.org>