tests: extend checkSystemCondition to match for strings or numbers
#14399
In essence the state values are taken from the vdc database. The index
of each value is not ready accessible via an API, that is exposed to
addons. E.g. a vdc might have states heating / cooking / defrost and
nacho-soup. Creating a filter for state cooking we must lookup the index,
probably 2 since 0 is the implicit 'not-set' value.
This patch is bypassing the lookup of the state-index, instead it allows
to match directly against the string value. For legacy reason we must
preserve integer matches, since 1(=active)/2(=inactive) are hard coded
in too many places. This introduces ambiguity if strings values encode
integers, e.g.
state-desc = "1"
state-index = 5
A filter matching the value '1' will now match the description "1"
while the filter probably should have matched the index 1(=active)

Signed-off-by: Andreas Fenkart <andreas.fenkart@dev.digitalstrom.org>