Replace hard-coded number by SensorIDUnknownType
Hard coded numbers are bad. Since the number is not registered
in modelconst it might accidentally be assumed to be free to use.
SensorIDUnknownType is used to set sensorType to a known invalid
state, in case parsing user data fails, we can verify that parsing
failed by comparing against this value

Signed-off-by: Andreas Fenkart <andreas.fenkart@dev.digitalstrom.org>