Use pkg-config to find librrd
Yocto 2.2 uses rrdtool 1.6.0 which is now threadsafe by default, so
-lrrd_th is now deprecated, this library no longer exists.

I'll update our pre 2.2 OE with rrdtool 1.6.0 so that we do not run into
build problems once this change is merged.

At the same time we will try to link to the -lrrd_th version which will
only succeed if the librrd version that is installed is < than 1.6.0,
this will be mostly relevant for PC builds.