structuremanipulater: replace zone.getName by intToString(zone.getID)
Thanks to the formidable implementation, zone.getName is the empty
string by default and not zone + intToString(zoneID) if not set
otherwise.
This assumed behavior can not be implemented without changing
base classes (virtual getName) or copy-paste of to much code
(utf8 conversion + setModelDirty)
Reverting back to intToString(getID) yippiee!

Signed-off-by: Andreas Fenkart <andreas.fenkart@dev.digitalstrom.org>