Do not overwrite pairing information of BL-SK204
Unfortunately the SK-204 firmware does not provide the correct number of
paired dSUIDs, so we can't use the mechanism where we read this
value from the device. As a workaround we forcefully set the correct value in
the dSS model.

Since there is some generic code for devices that actually supports
the dynamic readout and since this particular information is only a subset of
an otherwise valid configuration, we need to make sure that the readout
will not overwrite our hardcoded values for the paired devices setting.

So, make sure the paired devices value that was read out from the device
configuration is not applied to SK-204.

implements #15331