Fix setting of paired devices in device class
This was somewhat a chicken and an egg problem. The setPairedDevices()
function is receiving a new number of paired devices as parameter. Before
applying this value to the m_pairedDevices member variable it does some checks.

One of these checks includes a call to isMainDevice(), which does some
checks of its own. However, isMainDevice() uses m_pairedDevices in order
to perform those checks, and in the above case it uses old information,
because m_pairedDevices has not yet been set to the new value.

This leads to a wrong calculation in isMainDevice() which then causes
wrong assumptions in m_pairedDevices.

implements #15331