system_handlers: replace ::setup by constructor
Most plugins create a copy of the dSS::Event then execute blocking
calls in the context of their private task handler. This mechanism
was probably invented because some plugins execute blocking calls (http),
that would block the main loop for too long. (but some just copied the
paradigm but don't actually issue blocking calls)
When using exceptions we can copy the event fields in the constructor.
::setup is always called after constructor and before adding event
to local task queue

Signed-off-by: Andreas Fenkart <andreas.fenkart@dev.digitalstrom.org>