handler: bug: introduce ActionExecute
fixes #16943
Wrap the 99% of the implementation of SystemEvenActionExecute into
new class ActionExecute, so it can be reused by HighlevelEvent
without the remaining 1% logic in the constructor.
ActionExecute needs a PropertyTree node to get parameters to create
a new event. While SystemActionEventExecute expects the 'path' to be a
field in the event, SystemEventHighlevel expects an 'id' in the event
then generates the path according to a format.
SystemEventActionExecute checks in its constructor that the event
contains 'path', and throws and exception if not. SystemEventHighlevel
events do not contain that field, hence can not subclass
SystemEventActionExecute. By factoring out the implementation
into ActionExecute, both can share the code using delegation instead of
inheritance.

Signed-off-by: Andreas Fenkart <andreas.fenkart@dev.digitalstrom.org>