set setButtonGroupMembership() before setting active group
if the button is set to Area setDeviceButtonActiveGroup() also resets
the mode to room button. In the same call also the m_buttonGroupMembership
is written down to the device. The dSM will automatically add a device to
the group in m_buttonGroupMembership. Therefore, we mus make sure that we
do not write the previous group here.

refs #17343